+++ crates/aptos-dkg/src/pvss/chunky/chunks.rs
@@ -5,7 +5,7 @@ use ark_ff::{BigInteger, PrimeField};
 
 /// Converts a field element into little-endian chunks of `num_bits` bits.
 /// Supports any chunk size from 1 to 64 bits (e.g. 43 or 52 bits). Made `pub` for tests.
-pub fn scalar_to_le_chunks<F: PrimeField>(num_bits: u8, scalar: &F) -> Vec<F> {
+pub fn scalar_to_le_chunks<F: PrimeField>(num_bits: usize, scalar: &F) -> Vec<F> {
     assert!(num_bits > 0 && num_bits <= 64, "Invalid chunk size");
     if num_bits.is_multiple_of(8) {
         return scalar_to_le_chunks_byte_aligned(num_bits, scalar);
@@ -39,7 +39,7 @@ pub fn scalar_to_le_chunks<F: PrimeField>(num_bits: u8, scalar: &F) -> Vec<F> {
 /// Faster for those sizes since it works on whole bytes. Made `pub` for tests and benchmarks.
 ///
 /// Benchmarks suggest it's 1.5x faster than `scalar_to_le_chunks` for 32 bits, and 2x faster than `scalar_to_le_chunks` for 64 bits.
-pub fn scalar_to_le_chunks_byte_aligned<F: PrimeField>(num_bits: u8, scalar: &F) -> Vec<F> {
+pub fn scalar_to_le_chunks_byte_aligned<F: PrimeField>(num_bits: usize, scalar: &F) -> Vec<F> {
     assert!(
         num_bits.is_multiple_of(8) && num_bits > 0 && num_bits <= 64,
         "Invalid chunk size (must be multiple of 8)"
@@ -66,7 +66,7 @@ pub fn scalar_to_le_chunks_byte_aligned<F: PrimeField>(num_bits: u8, scalar: &F)
 /// Reconstructs a field element from `num_bits`-bit chunks (little-endian order). Made `pub` for tests
 /// This is the inverse of `scalar_to_le_chunks`.
 /// Chunks must be in the range [0, 2^num_bits).
-pub fn le_chunks_to_scalar<F: PrimeField>(num_bits: u8, chunks: &[F]) -> F {
+pub fn le_chunks_to_scalar<F: PrimeField>(num_bits: usize, chunks: &[F]) -> F {
     assert!(num_bits > 0 && num_bits <= 64, "Invalid chunk size");
 
     let base = F::from(1u128 << num_bits); // need u128 in the case where `num_bits` is 64, because of `chunk * multiplier`
@@ -112,7 +112,7 @@ mod tests {
     #[test]
     fn test_zero_roundtrips() {
         let zero = Fr::zero();
-        for &num_bits in &[1u8, 7, 8, 16, 32, 52, 64] {
+        for &num_bits in &[1, 7, 8, 16, 32, 52, 64] {
             let chunks = scalar_to_le_chunks(num_bits, &zero);
             let reconstructed = le_chunks_to_scalar(num_bits, &chunks);
             assert_eq!(
@@ -127,7 +127,7 @@ mod tests {
     #[test]
     fn test_byte_aligned_matches_arbitrary() {
         let mut rng = test_rng();
-        for &num_bits in &[8u8, 16, 32, 64] {
+        for &num_bits in &[8, 16, 32, 64] {
             for _ in 0..50 {
                 let scalar: Fr = Fr::rand(&mut rng);
                 let chunks_arb = scalar_to_le_chunks(num_bits, &scalar);
